backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Pillow==10.0.0
piexif==1.1.3
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
